pic: compiler compat fixes.
sdcc was unable to find my stdbool due to a misspelling of a var in
sdcc.mk (sdcc has a --print-search-dirs option that assisted in debuging
this).

I split the header files (stdbool.h and stdint.h) between mcc18 and sdcc
so I could override sdcc's stdbool.h but not its stdint.h (which is
decent).

The include hack for mcc18's stdbool was due to me not wanting to add
another path to the mcc18 search path. The way mcc18.mk is writen, a
great deal of work is needed to get one path (due to unix/windows
directory conversions). Need to rewrite it so I can have one var full of
windows paths, and then deal with it all at once.