Modified the public interface and header comments
Everything that is for internal use only was removed from hax.h to make it more
friendly to skimming. Similarly, the documentation was modified to to be inline 
to increase the readability of the file and force it to fit on a reasonably
sized terminal with no scrolling. These changes, unfortunately, will break a 
great deal of Hax's internals.