Fixed stm32loader's buffering in Mac OS X
The open-source usb-to-serial driver for the Prolific PL2303 buffers data even
if there is no process reading from the file descriptor. This causes
stm32loader to treat the buffered data as the microcontroller's real-time
responses and abort the transfer. This was fixed by clearing the input and
output (just in case...) buffers before initializing the transfer. This,
unfortunately, does not fix the verification step.